Parsing javascript quotes like this:       "('adding '+a+' and ' +b);\    // separate string using "\"
Starts with a " and has a " at the end of the line, though that is part of a
comment and I assume the \ after the ; means it is a line continuation
indicator and the rest of the line should be ignored.  So, I have updated
parseString() to do just that.     

This correctly parses:
    Test\jsFunc_tutorial.js